done on g++

+++ WordTree.hpp
@@ -25,6 +25,6 @@ class WordTree
     std::size_t size();
 
   private:
-    bool findRecursive(std::string word, std::shared_ptr<TreeNode> currNode, int i);
+    bool findRecursive(std::string word, std::shared_ptr<TreeNode> currNode, long unsigned int i);
     std::shared_ptr<TreeNode> getMyNode(std::string partial, std::shared_ptr<TreeNode> currNode, long unsigned int i);
 };
